Added in timing code to time each sampling and get a total run time

+++ src/renderer.h
@@ -64,7 +64,7 @@ struct renderer {
 
 
  //Renderer
-int renderThread();
+double renderThread();
 
  /*
 //Renderer
